requirements.txt
google-genai>=0.8
numpy
python-dotenv
